Fix Id generation using generated id only

+++ src/main/java/com/uca/m2/pdd/util/FieldUtils.java
@@ -3,10 +3,10 @@ package com.uca.m2.pdd.util;
 public class FieldUtils {
     public static final int DEFAULT_STRING_MIN_SIZE = 1;
     public static final int DEFAULT_STRING_MAX_SIZE = 255;
-    public static final int DEFAULT_LONGITUDE_MIN_SIZE = -180;
-    public static final int DEFAULT_LONGITUDE_MAX_SIZE = 180;
-    public static final int DEFAULT_LATITUDE_MIN_SIZE = -90;
-    public static final int DEFAULT_LATITUDE_MAX_SIZE = 90;
+    public static final double DEFAULT_LONGITUDE_MIN_SIZE = -180;
+    public static final double DEFAULT_LONGITUDE_MAX_SIZE = 180;
+    public static final double DEFAULT_LATITUDE_MIN_SIZE = -90;
+    public static final double DEFAULT_LATITUDE_MAX_SIZE = 90;
     public static final String REGEX_PHONE_NUMBER_VALIDATION = "^[0-9-+]*$";
     public static final String ERROR_MESSAGE_PHONE_NUMBER_VALIDATION = "Invalid phone number";
     public static final String REGEX_PASSWORD_VALIDATION = "^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[@#$%^&+=])(?=\\S+$).{8,}$";
